+++ nym-vpn-apple/Home/Sources/Home/HomeViewModel.swift
@@ -15,15 +15,14 @@ import TunnelMixnet
 import TunnelStatus
 import Tunnels
 import UIComponents
-#if os(iOS)
 import ImpactGenerator
-#elseif os(macOS)
+#if os(macOS)
 import GRPCManager
 import HelperInstall
 import HelperManager
 #endif
 
-public class HomeViewModel: HomeFlowState {
+@MainActor public class HomeViewModel: HomeFlowState {
     let title = "NymVPN".localizedString
     let connectToLocalizedTitle = "connectTo".localizedString
     let networkSelectLocalizedTitle = "selectNetwork".localizedString
@@ -34,15 +33,22 @@ public class HomeViewModel: HomeFlowState {
     let externalLinkManager: ExternalLinkManager
     let gatewayManager: GatewayManager
     let networkMonitor: NetworkMonitor
-#if os(iOS)
     let impactGenerator: ImpactGenerator
-#elseif os(macOS)
+#if os(macOS)
     let grpcManager: GRPCManager
     let helperManager: HelperManager
 #endif
     let messagesManager: MessagesManager
-    let anonymousButtonViewModel = NetworkButtonViewModel(type: .mixnet5hop)
-    let fastButtonViewModel = NetworkButtonViewModel(type: .wireguard)
+    let anonymousButtonViewModel = NetworkButtonViewModel(
+        type: .mixnet5hop,
+        appSettings: .shared,
+        connectionManager: .shared
+    )
+    let fastButtonViewModel = NetworkButtonViewModel(
+        type: .wireguard,
+        appSettings: .shared,
+        connectionManager: .shared
+    )
 
     @ObservedObject var connectionManager: ConnectionManager
     var cancellables = Set<AnyCancellable>()
@@ -102,15 +108,15 @@ public class HomeViewModel: HomeFlowState {
 
 #if os(iOS)
     public init(
-        appSettings: AppSettings = .shared,
-        connectionManager: ConnectionManager = .shared,
-        configurationManager: ConfigurationManager = .shared,
-        credentialsManager: CredentialsManager = .shared,
-        networkMonitor: NetworkMonitor = .shared,
-        externalLinkManager: ExternalLinkManager = .shared,
-        gatewayManager: GatewayManager = .shared,
-        impactGenerator: ImpactGenerator = .shared,
-        messagesManager: MessagesManager = .shared
+        appSettings: AppSettings,
+        connectionManager: ConnectionManager,
+        configurationManager: ConfigurationManager,
+        credentialsManager: CredentialsManager,
+        networkMonitor: NetworkMonitor,
+        externalLinkManager: ExternalLinkManager,
+        gatewayManager: GatewayManager,
+        impactGenerator: ImpactGenerator,
+        messagesManager: MessagesManager
     ) {
         self.appSettings = appSettings
         self.connectionManager = connectionManager
@@ -127,16 +133,17 @@ public class HomeViewModel: HomeFlowState {
     }
 #elseif os(macOS)
     public init(
-        appSettings: AppSettings = .shared,
-        connectionManager: ConnectionManager = .shared,
-        configurationManager: ConfigurationManager = .shared,
-        credentialsManager: CredentialsManager = .shared,
-        networkMonitor: NetworkMonitor = .shared,
-        grpcManager: GRPCManager = .shared,
-        helperManager: HelperManager = .shared,
-        externalLinkManager: ExternalLinkManager = .shared,
-        gatewayManager: GatewayManager = .shared,
-        messagesManager: MessagesManager = .shared
+        appSettings: AppSettings,
+        connectionManager: ConnectionManager,
+        configurationManager: ConfigurationManager,
+        credentialsManager: CredentialsManager,
+        networkMonitor: NetworkMonitor,
+        grpcManager: GRPCManager,
+        helperManager: HelperManager,
+        externalLinkManager: ExternalLinkManager,
+        gatewayManager: GatewayManager,
+        impactGenerator: ImpactGenerator,
+        messagesManager: MessagesManager
     ) {
         self.appSettings = appSettings
         self.connectionManager = connectionManager
@@ -147,6 +154,7 @@ public class HomeViewModel: HomeFlowState {
         self.grpcManager = grpcManager
         self.helperManager = helperManager
         self.externalLinkManager = externalLinkManager
+        self.impactGenerator = impactGenerator
         self.gatewayManager = gatewayManager
         self.messagesManager = messagesManager
         super.init()
@@ -154,10 +162,6 @@ public class HomeViewModel: HomeFlowState {
         setup()
     }
 #endif
-
-    deinit {
-        cancellables.forEach { $0.cancel() }
-    }
 }
 
 // MARK: - Navigation -
@@ -168,10 +172,12 @@ public extension HomeViewModel {
     }
 
     @MainActor func navigateToEntryGateways() {
+        impactGenerator.softImpact()
         path.append(HomeLink.entryGateways)
     }
 
     @MainActor func navigateToExitGateways() {
+        impactGenerator.softImpact()
         path.append(HomeLink.exitGateways)
     }