+++ nym-vpn-apple/Settings/Sources/Settings/Support/ResetVPNProfileDialogViewModel.swift
@@ -1,12 +1,8 @@
 import SwiftUI
-#if os(iOS)
 import ImpactGenerator
-#endif
 
-final class ResetVPNProfileDialogViewModel: ObservableObject {
-#if os(iOS)
+@MainActor final class ResetVPNProfileDialogViewModel: ObservableObject {
     let impactGenerator: ImpactGenerator
-#endif
     let resetVpnProfileTitle = "settings.resetVpnProfileTitle".localizedString
     let resetVpnProfileSubtitle = "settings.resetVpnProfileSubtitle".localizedString
     let yesLocalizedString = "logs.yes".localizedString
@@ -16,21 +12,13 @@ final class ResetVPNProfileDialogViewModel: ObservableObject {
 
     @Binding var isDisplayed: Bool
 
-#if os(iOS)
     init(
         isDisplayed: Binding<Bool>,
-        impactGenerator: ImpactGenerator = ImpactGenerator.shared,
+        impactGenerator: ImpactGenerator,
         action: @escaping () -> Void
     ) {
         _isDisplayed = isDisplayed
         self.impactGenerator = impactGenerator
         self.action = action
     }
-#endif
-#if os(macOS)
-    init(isDisplayed: Binding<Bool>, action: @escaping () -> Void) {
-        _isDisplayed = isDisplayed
-        self.action = action
-    }
-#endif
 }