+++ nym-vpn-apple/CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.11.0]
+
+### Added
+
+- US region selection
+
+### Fixed
+
+- Country loosing selection state 
+
+## [2.10.0] - 2025-10-02
 
 ### Added